Ajustes da interface do programa

+++ manga_dex.py
@@ -1,6 +1,6 @@
 import requests
 import os
-from alive_progress import alive_bar
+import alive_progress 
 import sys
 
 def remover_repetidos(capitulos:list):
@@ -26,6 +26,9 @@ def baixar_capitulos(capitulos, inicio, fim):
         num_chap = capitulos[i]['Num_Capitulo']
         vol_chap = capitulos[i]['Volume']
         
+        if vol_chap == None:
+            vol_chap = "Nenhum"
+        
         chap = requests.get(f"{base_url}/at-home/server/{chap_id}")
         
         r_json = chap.json()
@@ -36,19 +39,26 @@ def baixar_capitulos(capitulos, inicio, fim):
         host = r_json["baseUrl"]
         chapter_hash = r_json["chapter"]["hash"]
         data_saver = r_json["chapter"]["dataSaver"]
+        data = r_json["chapter"]["data"]
 
-        
-        with alive_bar(len(data_saver), title = f"Capitulo {num_chap}") as bar:
+        with alive_progress.alive_bar(len(data_saver), title = f"Capitulo {num_chap} - Vol {vol_chap}") as bar:
             for index, page in enumerate(data_saver):
+                if os.path.exists(f"{folder_path}/Page {index}"):
+                    continue
+                r = requests.get(f"{host}/data/{chapter_hash}/{page}")
+                if r.status_code == 404:
+                    page = data[index]
                     r = requests.get(f"{host}/data/{chapter_hash}/{page}")
-
-                    with open(f"{folder_path}/Page {index}", mode="wb") as f:
-                        f.write(r.content)
-                    bar()
+                with open(f"{folder_path}/Page {index}", mode="wb") as f:
+                    f.write(r.content)
+                bar()
+            #print("Cápitulo já existente")
 
 base_url = "https://api.mangadex.org"
 
 languages =['pt-br']
+
+print("\nIniciando conexão com MangaDex\n")
 
 title = sys.argv[1]
 
@@ -65,8 +75,9 @@ if len(mangas_achados) != 0:
         title = m["attributes"]['title']['en']
         print(f"[{index + 1}] - {title}")
 
-    resp = int(input("Qual mangá deseja selecionar? ")) - 1
-
+    resp = int(input("\nQual mangá deseja selecionar? ")) - 1
+    print("\n")
+    
     manga_id = mangas_achados[resp]['id']
     nome_manga = mangas_achados[resp]["attributes"]['title']['en']
 
@@ -94,15 +105,19 @@ if len(mangas_achados) != 0:
     capitulos_sem_repeticao = remover_repetidos(capitulos_organizados)
 
     for index, cap_vol in enumerate(capitulos_sem_repeticao):
-        print(f"({index}) - Capitulo {cap_vol['Num_Capitulo']} - {cap_vol['Titulo']}")
+        print(f"({index + 1}) - Capitulo {cap_vol['Num_Capitulo']} - {cap_vol['Titulo']}")
 
     #* Ajustar para poder escolher os capitulos para baixar
+
+    print(f"\n{len(capitulos_sem_repeticao)} Capitulos encontrados\n")
 
     escolha_cap = input("Quais capitulos deseja baixar? ")
 
     if "-" in escolha_cap:
         cap_ini, cap_fim = escolha_cap.split("-")
-        baixar_capitulos(capitulos_sem_repeticao, int(cap_ini), int(cap_fim))
+        baixar_capitulos(capitulos_sem_repeticao, int(cap_ini) - 1, int(cap_fim) - 1)
+    elif escolha_cap.isnumeric():
+        baixar_capitulos(capitulos_sem_repeticao, int(escolha_cap) - 1, int(escolha_cap) - 1)
     elif escolha_cap == "todos":
         baixar_capitulos(capitulos_sem_repeticao, 0, len(capitulos_sem_repeticao))
     elif escolha_cap == "nenhum":
